fix: api > endpoints 중복 함수 삭제 (#160)
- `playlist.ts`에 중복 함수 삭제
- 관심사에 따라 조회 / 수정, 삭제, 추가로 파일 분리
