+++ src/api/endpoints/comment.ts
@@ -1,3 +1,5 @@
+// 댓글 관련 파이어베이스 API 함수
+
 import {
   getFirestore,
   collection,
@@ -13,18 +15,18 @@ import {
   deleteDoc,
 } from 'firebase/firestore';
 
-import { app } from '@/api'; // Firebase 앱 초기화 파일
 import { Comment } from '@/types/playlist';
 
-const db = getFirestore(app);
+const db = getFirestore();
 
+// 해당 플레이리스트 댓글 조회
 export const getPlaylistComments = async (
   playlistId: string | undefined,
-  limitCount: number = 10
+  limitCount: number = 20
 ): Promise<Comment[]> => {
   try {
     if (!playlistId) {
-      throw new Error('Playlist ID is required');
+      throw new Error('Playlist ID가 필요합니다.');
     }
 
     const commentsRef = collection(db, 'comments');
@@ -50,11 +52,12 @@ export const getPlaylistComments = async (
       };
     });
   } catch (error) {
-    console.error(`Error fetching comments for playlist ${playlistId}:`, error);
+    console.error(`getPlaylistComments의 ${playlistId}에 대한 에러가 발생했습니다.:`, error);
     return [];
   }
 };
 
+// 댓글 추가
 export const addComment = async (
   playlistId: string,
   userId: string,
@@ -85,24 +88,23 @@ export const addComment = async (
 
     return commentId;
   } catch (error) {
-    console.error('Error adding comment:', error);
     throw new Error('댓글 추가에 실패했습니다.');
   }
 };
 
+// 댓글 삭제
 export const deleteComment = async (playlistId: string, commentId: string): Promise<string> => {
   try {
-    const commentsRef = doc(db, 'comments', commentId); //comments의 특정 commentId를 가진 문서를 저장
+    const commentsRef = doc(db, 'comments', commentId); // comments의 특정 commentId를 가진 문서를 저장
     await deleteDoc(commentsRef); // 특정 commentId를 가진 문서(댓글) 삭제
 
     const playlistRef = doc(db, 'playlists', playlistId); // playlists의 특정 playlistId를 가진 문서를 저장
     await updateDoc(playlistRef, {
-      commentCount: increment(-1), //해당 문서(playlistRef)의 commentCount에 -1해서 문서 상태를 업데이트
+      commentCount: increment(-1), // 해당 문서(playlistRef)의 commentCount에 -1해서 문서 상태를 업데이트
     });
 
     return commentId;
   } catch (error) {
-    console.error('Error adding comment:', error);
     throw new Error('댓글 삭제에 실패했습니다.');
   }
 };
