+++ src/api/endpoints/fork.ts
@@ -1,34 +1,34 @@
+// 플레이리스트 포크(구독) 관련 파이어베이스 API 함수
+
 import { getFirestore, runTransaction, doc, getDoc } from 'firebase/firestore';
 
-import { app } from '@/api'; // Firebase 앱 초기화 파일
+const db = getFirestore();
 
-const db = getFirestore(app);
-
-// 초기 구독 상태 확인 함수
-export const getInitialForkedState = async (
-  userId: string,
-  playlistId: string
-): Promise<boolean> => {
+// 사용자가 해당 플레이리스트를 포크(구독)했는지 확인
+export const getIsForkedState = async (userId: string, playlistId: string): Promise<boolean> => {
   const userPlaylistRef = doc(db, 'userPlaylists', userId);
   const userPlaylistSnap = await getDoc(userPlaylistRef);
-  if (userPlaylistSnap.exists()) {
-    const userData = userPlaylistSnap.data();
-    return userData.forked?.includes(playlistId) || false;
+
+  if (!userPlaylistSnap.exists()) {
+    return false;
   }
-  return false;
+
+  const userPlaylist = userPlaylistSnap.data();
+  return userPlaylist.forked?.includes(playlistId) || false;
 };
 
-// 구독 토글 함수
-export const toggleFork = async (
+// 포크(구독) 토글 버튼 처리
+// 포크가 되어있다면, forked 배열에서 해당 플레이리스트 ID 삭제 / 플레이리스트 포크 수(forkCount) -1
+// 포크가 안 되어있다면, forked 배열에 해당 플레이리스트 ID 추가 / 플레이리스트 포크 수(forkCount) +1
+export const toggleForkPlaylist = async (
   playlistId: string,
   userId: string,
   isCurrentlyForked: boolean
 ) => {
   try {
     const newForkState = await runTransaction(db, async (transaction) => {
-      const userPlaylistRef = doc(db, 'userPlaylists', userId);
-
-      const playlistRef = doc(db, 'playlists', playlistId);
+      const userPlaylistRef = doc(db, 'userPlaylists', userId); // 사용자와 플레이리스트 관계를 알 수 있는 컬렉션
+      const playlistRef = doc(db, 'playlists', playlistId); // 플레이리스트 컬렉션
 
       const [playlistDoc, userPlaylistDoc] = await Promise.all([
         transaction.get(playlistRef),
@@ -36,7 +36,7 @@ export const toggleFork = async (
       ]);
 
       if (!playlistDoc.exists()) {
-        throw new Error('User playlist document not found');
+        throw new Error('플레이리스트가 없습니다.');
       }
 
       const playlistData = playlistDoc.data();
@@ -44,17 +44,17 @@ export const toggleFork = async (
       const forkedPlaylists = userPlaylistData.forked || [];
 
       if (isCurrentlyForked) {
-        // 구독 취소
+        // 포크(구독) 취소
         transaction.update(userPlaylistRef, {
-          forked: forkedPlaylists.filter((id: string) => id !== playlistId),
+          forked: forkedPlaylists.filter((id: string) => id !== playlistId), // 플레이리스트 ID 삭제
         });
         transaction.update(playlistRef, {
           forkCount: (playlistData.forkCount || 0) - 1,
         });
       } else {
-        // 구독
+        // 포크(구독)
         transaction.update(userPlaylistRef, {
-          forked: [...forkedPlaylists, playlistId],
+          forked: [...forkedPlaylists, playlistId], // 플레이리스트 ID 추가
         });
         transaction.update(playlistRef, {
           forkCount: (playlistData.forkCount || 0) + 1,
@@ -65,7 +65,7 @@ export const toggleFork = async (
     });
     return newForkState;
   } catch (error) {
-    console.error('Error toggling fork:', error);
+    console.error('toggleForkPlaylist() 에러가 발생했습니다: ', error);
     throw error;
   }
 };
