+++ src/api/endpoints/like.ts
@@ -1,31 +1,34 @@
+// 플레이리스트 좋아요 관련 파이어베이스 API 함수
+
 import { getFirestore, runTransaction, doc, getDoc } from 'firebase/firestore';
 
 const db = getFirestore();
 
-// isLiked 상태 초기화 함수
-export const getInitialLikedState = async (
-  userId: string,
-  playlistId: string
-): Promise<boolean> => {
+// 사용자가 해당 플레이리스트를 좋아요했는지 확인
+export const getIsLikedState = async (userId: string, playlistId: string): Promise<boolean> => {
   const userPlaylistRef = doc(db, 'userPlaylists', userId);
   const userPlaylistSnap = await getDoc(userPlaylistRef);
-  if (userPlaylistSnap.exists()) {
-    const userData = userPlaylistSnap.data();
-    return userData.liked?.includes(playlistId) || false;
+
+  if (!userPlaylistSnap.exists()) {
+    return false;
   }
-  return false;
+
+  const userPlaylist = userPlaylistSnap.data();
+  return userPlaylist.liked?.includes(playlistId) || false;
 };
 
-// 좋아요 토글 함수
-export const togglePlaylistLike = async (
+// 좋아요 토글 버튼 처리
+// 좋아요가 되어있다면, liked 배열에서 해당 플레이리스트 ID 삭제 / 플레이리스트 좋아요 수(likeCount) -1
+// 좋아요가 안 되어있다면, liked 배열에 해당 플레이리스트 ID 추가 / 플레이리스트 좋아요 수(likeCount) +1
+export const toggleLikePlaylist = async (
   playlistId: string,
   userId: string,
   isCurrentlyLiked: boolean
 ) => {
   try {
     const newLikeState = await runTransaction(db, async (transaction) => {
-      const userPlaylistRef = doc(db, 'userPlaylists', userId);
-      const playlistRef = doc(db, 'playlists', playlistId);
+      const userPlaylistRef = doc(db, 'userPlaylists', userId); // 사용자와 플레이리스트 관계를 알 수 있는 컬렉션
+      const playlistRef = doc(db, 'playlists', playlistId); // 플레이리스트 컬렉션
 
       const [playlistDoc, userPlaylistDoc] = await Promise.all([
         transaction.get(playlistRef),
@@ -33,7 +36,7 @@ export const togglePlaylistLike = async (
       ]);
 
       if (!playlistDoc.exists()) {
-        throw new Error('Playlist not found');
+        throw new Error('플레이리스트가 없습니다.');
       }
 
       const playlistData = playlistDoc.data();
@@ -43,7 +46,7 @@ export const togglePlaylistLike = async (
       if (isCurrentlyLiked) {
         // 좋아요 취소
         transaction.update(userPlaylistRef, {
-          liked: likedPlaylists.filter((id: string) => id !== playlistId),
+          liked: likedPlaylists.filter((id: string) => id !== playlistId), // 플레이리스트 ID 삭제
         });
         transaction.update(playlistRef, {
           likeCount: (playlistData.likeCount || 0) - 1,
@@ -51,7 +54,7 @@ export const togglePlaylistLike = async (
       } else {
         // 좋아요
         transaction.update(userPlaylistRef, {
-          liked: [...likedPlaylists, playlistId],
+          liked: [...likedPlaylists, playlistId], // 플레이리스트 ID 추가
         });
         transaction.update(playlistRef, {
           likeCount: (playlistData.likeCount || 0) + 1,
@@ -63,7 +66,7 @@ export const togglePlaylistLike = async (
 
     return newLikeState;
   } catch (error) {
-    console.error('Error toggling like:', error);
+    console.error('toggleLikePlaylist() 에러가 발생했습니다: ', error);
     throw error;
   }
 };
