+++ src/api/endpoints/user.ts
@@ -1,32 +1,18 @@
+// 사용자 관련 파이어베이스 API 함수
+
 import { getFirestore, doc, getDoc } from 'firebase/firestore';
 
 import { app } from '@/api'; // Firebase 앱 초기화 파일
-import { UserModel } from '@/types/user';
+import { UserModel, UserPlaylistsModel } from '@/types/user';
 
 const db = getFirestore(app);
 
-export interface User {
-  id: string;
-  username: string;
-  userBio: string;
-  email: string;
-  profileImg: string;
-  playlistCount: number;
-  totalLikes: number;
-  totalForks: number;
-}
-
-export interface UserPlaylists {
-  created: string[];
-  forked: string[];
-  liked: string[];
-}
-
+// 사용자 정보 조회
 export const getUserData = async (userId: string): Promise<UserModel | null> => {
   if (!userId) {
-    console.error('유효한 사용자 ID가 없습니다.');
-    return null;
+    throw new Error('유효한 사용자 ID가 없습니다.');
   }
+
   try {
     const userDocRef = doc(db, 'users', userId);
     const userDoc = await getDoc(userDocRef);
@@ -48,12 +34,12 @@ export const getUserData = async (userId: string): Promise<UserModel | null> => 
       totalForks: data.totalForks || 0,
     };
   } catch (error) {
-    console.error('Error fetching user data:', error);
-    return null;
+    throw new Error('getUserData() 에러가 발생했습니다.');
   }
 };
 
-export const getUserPlaylists = async (userId: string): Promise<UserPlaylists | null> => {
+// 사용자와 플레이리스트와의 관계(생성, 좋아요, 포크) 조회
+export const getUserPlaylists = async (userId: string): Promise<UserPlaylistsModel | null> => {
   try {
     const userPlaylistsRef = doc(db, 'userPlaylists', userId);
     const userPlaylistsDoc = await getDoc(userPlaylistsRef);
@@ -62,14 +48,14 @@ export const getUserPlaylists = async (userId: string): Promise<UserPlaylists | 
       return null;
     }
 
-    const data = userPlaylistsDoc.data() as UserPlaylists;
+    const data = userPlaylistsDoc.data() as UserPlaylistsModel;
     return {
+      userId: data.userId || '',
       created: data.created || [],
       forked: data.forked || [],
       liked: data.liked || [],
     };
   } catch (error) {
-    console.error('Error fetching user playlists:', error);
-    return null;
+    throw new Error('getUserPlaylists() 에러가 발생했습니다.');
   }
 };
