+++ src/api/endpoints/youtube.ts
@@ -1,3 +1,5 @@
+// Youtube API
+
 import { fetchYoutubeData } from '@/api/apiClient';
 import { getVideoId } from '@/utils/getVideoId';
 
@@ -13,21 +15,12 @@ export const getVideoData = async (url: string) => {
   const item = response.data.items[0];
 
   if (!item) {
-    throw new Error('No video data found');
+    throw new Error('해당 동영상을 찾을 수 없습니다.');
   }
 
   return {
     title: item.snippet.title,
-    // channelTitle: item.snippet.channelTitle,
-    // channelId: item.snippet.channelId,
     thumbnailUrl: item.snippet.thumbnails.medium.url,
     duration: item.contentDetails.duration,
-    // publishedAt: new Date(item.snippet.publishedAt).toLocaleDateString(),
   };
 };
-
-// title: 영상 제목
-// channelTitle: 채널 이름
-// channelId: 체널 id
-// thumbnailUrl: 썸네일 url
-// publishedAt: 업로드 날짜
