+++ src/components/page/playlistdetail/thumBoxDetail.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react';
 
 import { css } from '@emotion/react';
 
-import { getInitialLikedState, togglePlaylistLike } from '@/api/endpoints/like';
+import { getIsLikedState, toggleLikePlaylist } from '@/api/endpoints/like';
 import defaultThumbnail from '@/assets/images/default-thumb.svg';
 import CommentsButton from '@/components/common/buttons/CommentsButton';
 import LikesButton from '@/components/common/buttons/LikesButton';
@@ -45,7 +45,7 @@ const ThumbNailBoxDetail: React.FC<ThumbNailBoxDetailProps> = ({
   useEffect(() => {
     const fetchInitialLikedState = async () => {
       try {
-        const initialLikedState = await getInitialLikedState(userId, playlistId);
+        const initialLikedState = await getIsLikedState(userId, playlistId);
         setIsLiked(initialLikedState);
       } catch (error) {
         console.error('Error fetching initial liked state:', error);
@@ -63,7 +63,7 @@ const ThumbNailBoxDetail: React.FC<ThumbNailBoxDetailProps> = ({
     if (isLiked === null) return;
 
     try {
-      const newLikeState = await togglePlaylistLike(playlistId, userId, isLiked);
+      const newLikeState = await toggleLikePlaylist(playlistId, userId, isLiked);
       const newLikeCount = newLikeState ? localLikeCount + 1 : localLikeCount - 1;
       setIsLiked(newLikeState);
       setLocalLikeCount(newLikeCount);
