+++ src/constants/api.ts
@@ -6,3 +6,7 @@ export const firebaseConfig = {
   messagingSenderId: import.meta.env.VITE_FIREBASE_MESSAGING_SENDER_ID,
   appId: import.meta.env.VITE_FIREBASE_APP_ID,
 };
+
+export const BASE_URL = {
+  youtube: 'https://www.googleapis.com/youtube/v3',
+};
